[#200] Test: 캐시 테스트 전용 Redis 컨테이너 분리
캐시 테스트 2개만 Docker를 필요로 하도록 Redis 의존성을 분리

- RedisConfig: @ConditionalOnProperty(spring.cache.type=redis) 추가로 테스트 프로파일에서 비활성화
- IntegrationTestSupport: Redis 컨테이너 제거, NoOpShedLockConfig 추가
- RedisIntegrationTestSupport: 캐시 테스트 전용 Redis 컨테이너 싱글톤으로 분리
- CategoryCacheIntegrationTest / RefrigeratorCacheIntegrationTest: RedisIntegrationTestSupport 상속으로 변경
