+++ src/test/java/soon/fridgely/global/support/IntegrationTestSupport.java
@@ -1,34 +1,31 @@
 package soon.fridgely.global.support;
 
 import com.navercorp.fixturemonkey.FixtureMonkey;
+import net.javacrumbs.shedlock.core.LockProvider;
 import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.TestConfiguration;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Import;
+import org.springframework.context.annotation.Primary;
 import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.DynamicPropertyRegistry;
-import org.springframework.test.context.DynamicPropertySource;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.utility.DockerImageName;
+
+import java.util.Optional;
 
 @ActiveProfiles("test")
-@Import(TestSecurityConfig.class)
+@Import({TestSecurityConfig.class, IntegrationTestSupport.NoOpShedLockConfig.class})
 @TruncateTables
 @SpringBootTest
 public abstract class IntegrationTestSupport {
 
-    // Singleton Container: JVM 내 한 번만 시작 → Spring 컨텍스트 캐싱과 충돌 없음
-    static final GenericContainer<?> REDIS = new GenericContainer<>(DockerImageName.parse("redis:7-alpine"))
-        .withExposedPorts(6379);
-
-    static {
-        REDIS.start();
-    }
-
-    @DynamicPropertySource
-    static void redisProperties(DynamicPropertyRegistry registry) {
-        registry.add("spring.data.redis.host", REDIS::getHost);
-        registry.add("spring.data.redis.port", () -> REDIS.getMappedPort(6379));
-    }
-
     protected final FixtureMonkey fixtureMonkey = FixtureMonkeyFactory.get();
+
+    @TestConfiguration
+    static class NoOpShedLockConfig {
+        @Bean
+        @Primary
+        public LockProvider lockProvider() {
+            return lockConfiguration -> Optional.of(() -> {});
+        }
+    }
 
 }
